refactor : 각 컴포넌트, 별도의 파일로 분리
