+++ presentation/src/main/java/com/nexters/boolti/presentation/screen/refund/ReasonPage.kt
@@ -1,0 +1,70 @@
+package com.nexters.boolti.presentation.screen.refund
+
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalSoftwareKeyboardController
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import com.nexters.boolti.presentation.R
+import com.nexters.boolti.presentation.component.BTTextField
+import com.nexters.boolti.presentation.component.MainButton
+import com.nexters.boolti.presentation.theme.marginHorizontal
+import com.nexters.boolti.presentation.theme.point4
+
+@Composable
+fun ReasonPage(
+    reason: String,
+    onReasonChanged: (String) -> Unit,
+    onNextClick: () -> Unit,
+    modifier: Modifier = Modifier,
+) {
+    val keyboardController = LocalSoftwareKeyboardController.current
+    val interactionSource = remember { MutableInteractionSource() }
+
+    Column(
+        modifier = modifier.clickable(
+            interactionSource = interactionSource,
+            indication = null
+        ) {
+            keyboardController?.hide()
+        }
+    ) {
+        Text(
+            modifier = Modifier
+                .padding(top = 20.dp)
+                .padding(horizontal = marginHorizontal),
+            text = stringResource(id = R.string.refund_reason_label),
+            style = point4,
+        )
+        BTTextField(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(horizontal = marginHorizontal)
+                .height(160.dp)
+                .padding(top = 20.dp),
+            text = reason,
+            onValueChanged = onReasonChanged,
+            placeholder = stringResource(id = R.string.refund_reason_hint),
+        )
+
+        Spacer(modifier = Modifier.weight(1.0f))
+        MainButton(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(horizontal = marginHorizontal)
+                .padding(bottom = 8.dp),
+            onClick = onNextClick,
+            enabled = reason.isNotBlank(),
+            label = stringResource(id = R.string.next)
+        )
+    }
+}
